refactor: extract SeasonGroup from SeriesDetail.tsx
Move SeasonGroup function (lines 48-542) into its own file
frontend/src/components/series/SeasonGroup.tsx with a typed
SeasonGroupProps interface (all props readonly). Remove the 15+
imports from SeriesDetail that were exclusively used by SeasonGroup.
